+++ apps/rig_inbound_gateway/lib/rig_inbound_gateway/api_proxy/handler/kinesis.ex
@@ -19,7 +19,7 @@ defmodule RigInboundGateway.ApiProxy.Handler.Kinesis do
   def handle_http_request(conn, api, endpoint)
 
   @doc "CORS response for preflight request."
-  def handle_http_request(%{"method" => "OPTIONS"} = conn, _, %{"target" => "kinesis"}) do
+  def handle_http_request(%{method: "OPTIONS"} = conn, _, %{"target" => "kinesis"}) do
     conn
     |> with_cors()
     |> Conn.send_resp(:no_content, "")
@@ -30,18 +30,19 @@ defmodule RigInboundGateway.ApiProxy.Handler.Kinesis do
     %{params: %{"partition_key" => partition_key, "data" => data}} = conn
 
     kinesis_message =
-      Poison.encode!(%{
-        correlation_id: Codec.serialize(self()),
-        data: data,
+      data
+      |> Map.put("rig", %{
+        correlationID: Codec.serialize(self()),
         host: conn.host,
         method: conn.method,
-        request_path: conn.request_path,
+        requestPath: conn.request_path,
         port: conn.port,
-        remote_ip: conn.remote_ip,
-        req_headers: conn.req_headers,
+        remoteIP: to_string(:inet_parse.ntoa(conn.remote_ip)),
+        reqHeaders: Enum.map(conn.req_headers, &Tuple.to_list(&1)),
         scheme: conn.scheme,
-        query_string: conn.query_string
+        queryString: conn.query_string
       })
+      |> Poison.encode!()
 
     produce(
       _partition_key = partition_key,
@@ -99,6 +100,6 @@ defmodule RigInboundGateway.ApiProxy.Handler.Kinesis do
     conn
     |> Conn.put_resp_header("access-control-allow-origin", config().cors)
     |> Conn.put_resp_header("access-control-allow-methods", "*")
-    |> Conn.put_resp_header("access-control-allow-headers", "content-type")
+    |> Conn.put_resp_header("access-control-allow-headers", "content-type,authorization")
   end
 end