+++ apps/rig_inbound_gateway/test/rig_web/presence/channel_test.exs
@@ -14,86 +14,96 @@ defmodule RigInboundGatewayWeb.Presence.ChannelTest do
   end
 
   test "a user connecting to her own topic works" do
-    assert {:ok, _response, sock} = subscribe_and_join_user(
-      "testuser",
-      [@customer_role],
-      "user:testuser"
-    )
-    leave sock
+    assert {:ok, _response, sock} =
+             subscribe_and_join_user(
+               "testuser",
+               [@customer_role],
+               "user:testuser"
+             )
+
+    leave(sock)
   end
 
   test "a user connecting to someone else's topic with not authorized role fails" do
     fun = fn ->
-      assert {:error, _message} = subscribe_and_join_user(
-        "foo-user",
-        [@customer_role],
-        "user:bar-user"
-      )
+      assert {:error, _message} =
+               subscribe_and_join_user(
+                 "foo-user",
+                 [@customer_role],
+                 "user:bar-user"
+               )
     end
+
     assert capture_log(fun) =~ "unauthorized"
   end
 
   test "a user connecting to someone else's topic with authorized role works" do
-    assert {:ok, _response, sock} = subscribe_and_join_user(
-      "foo-user",
-      [@support_role],
-      "user:testuser"
-    )
-    leave sock
+    assert {:ok, _response, sock} =
+             subscribe_and_join_user(
+               "foo-user",
+               [@support_role],
+               "user:testuser"
+             )
+
+    leave(sock)
   end
 
   test "a user connecting to role specific topic with authorized role works" do
-    assert {:ok, _response, sock} = subscribe_and_join_user(
-      "foo-user",
-      [@support_role],
-      "role:customer"
-    )
-    leave sock
+    assert {:ok, _response, sock} =
+             subscribe_and_join_user(
+               "foo-user",
+               [@support_role],
+               "role:customer"
+             )
+
+    leave(sock)
   end
 
   test "a user connecting to role specific topic without authorized role fails" do
     fun = fn ->
-      assert {:error, _message} = subscribe_and_join_user(
-        "foo-user",
-        [@customer_role],
-        "role:customer"
-      )
+      assert {:error, _message} =
+               subscribe_and_join_user(
+                 "foo-user",
+                 [@customer_role],
+                 "role:customer"
+               )
     end
+
     assert capture_log(fun) =~ "unauthorized"
   end
 
-  @tag :smoke
-  test "subscribed user should receive message", %{conn: conn} do
-    true = RigOutboundGateway.Kafka.ready?("rig", 0, 10_000)
-    assert {:ok, _response, sock} = subscribe_and_join_user(
-      "testuser",
-      [@customer_role],
-      "user:testuser"
-    )
+  # @tag :smoke
+  # test "subscribed user should receive message", %{conn: conn} do
+  #   true = RigOutboundGateway.Kafka.ready?("rig", 0, 10_000)
+  #   assert {:ok, _response, sock} = subscribe_and_join_user(
+  #     "testuser",
+  #     [@customer_role],
+  #     "user:testuser"
+  #   )
 
-    body = ~s({"user":"testuser","foo":"bar"})
-    produce_kafka_message(conn, body)
+  #   body = ~s({"user":"testuser","foo":"bar"})
+  #   produce_kafka_message(conn, body)
 
-    expected_event = "message"
-    expected_payload = %{"user" => "testuser", "foo" => "bar"}
-    assert_broadcast ^expected_event, ^expected_payload, 3000
+  #   expected_event = "message"
+  #   expected_payload = %{"user" => "testuser", "foo" => "bar"}
+  #   assert_broadcast ^expected_event, ^expected_payload, 3000
 
-    leave sock
-  end
+  #   leave sock
+  # end
 
-  @tag :smoke
-  test "not subscribed user shouldn't receive message", %{conn: conn} do
-    true = RigOutboundGateway.Kafka.ready?("rig", 0, 10_000)
-    body = ~s({"user":"testuser","foo":"bar"})
-    produce_kafka_message(conn, body)
+  # @tag :smoke
+  # test "not subscribed user shouldn't receive message", %{conn: conn} do
+  #   true = RigOutboundGateway.Kafka.ready?("rig", 0, 10_000)
+  #   body = ~s({"user":"testuser","foo":"bar"})
+  #   produce_kafka_message(conn, body)
 
-    expected_event = "message"
-    expected_payload = %{"user" => "testuser", "foo" => "bar"}
-    refute_broadcast ^expected_event, ^expected_payload, 2000
-  end
+  #   expected_event = "message"
+  #   expected_payload = %{"user" => "testuser", "foo" => "bar"}
+  #   refute_broadcast ^expected_event, ^expected_payload, 2000
+  # end
 
   defp produce_kafka_message(conn, body) do
-    conn = post conn, "/kafka/produce", body
+    conn = post(conn, "/kafka/produce", body)
     assert conn.status == 200
   end
 end