+++ apps/rig_tests/test/proxy/response_from/kafka_test.exs
@@ -34,7 +34,16 @@ defmodule RigTests.Proxy.ResponseFrom.KafkaTest do
 
   setup do
     kafka_config = kafka_config()
-    {:ok, kafka_client} = RigKafka.start(kafka_config)
+
+    test_pid = self()
+
+    callback = fn
+      msg ->
+        send(test_pid, msg)
+        :ok
+    end
+
+    {:ok, kafka_client} = RigKafka.start(kafka_config, callback)
 
     on_exit(fn ->
       RigKafka.Client.stop_supervised(kafka_client)
@@ -57,12 +66,12 @@ defmodule RigTests.Proxy.ResponseFrom.KafkaTest do
     route(endpoint_path, fn %{query: query} ->
       correlation_id =
         query
-        |> Map.fetch!("correlation_id")
+        |> Map.fetch!("correlationID")
         |> URI.decode_www_form()
 
       message =
         async_response
-        |> Map.put("correlation_id", correlation_id)
+        |> Map.put("rig", %{"correlationID" => correlation_id})
         |> Jason.encode!()
 
       kafka_config = kafka_config()
@@ -114,5 +123,164 @@ defmodule RigTests.Proxy.ResponseFrom.KafkaTest do
     assert Jason.decode!(res_body) != sync_response
     # ...but the client got the response sent to the Kafka topic:
     assert Jason.decode!(res_body)["message"] == Map.fetch!(async_response, "message")
+  end
+
+  @tag :kafka
+  test "Given target is set to Kafka, the http OPTIONS request should handle CORS" do
+    test_name = "proxy-http-response-cors"
+
+    api_id = "mock-#{test_name}-api"
+    endpoint_id = "mock-#{test_name}-endpoint"
+    endpoint_path = "/#{endpoint_id}"
+
+    # We register the endpoint with the proxy:
+    rig_api_url = "http://localhost:#{@api_port}/v1/apis"
+    rig_proxy_url = "http://localhost:#{@proxy_port}"
+
+    body =
+      Jason.encode!(%{
+        id: api_id,
+        name: "Mock API",
+        version_data: %{
+          default: %{
+            endpoints: [
+              %{
+                id: endpoint_id,
+                type: "http",
+                not_secured: true,
+                method: "OPTIONS",
+                path: endpoint_path,
+                target: "kafka"
+              }
+            ]
+          }
+        },
+        proxy: %{
+          target_url: "http://localhost",
+          port: 3000
+        }
+      })
+
+    headers = [{"content-type", "application/json"}]
+    HTTPoison.post!(rig_api_url, body, headers)
+
+    # The client calls the proxy endpoint:
+    request_url = rig_proxy_url <> endpoint_path
+
+    %HTTPoison.Response{status_code: res_status, body: res_body, headers: res_headers} =
+      HTTPoison.options!(request_url, headers)
+
+    assert res_status == 204
+    assert res_body == ""
+
+    # CORS should verify origin
+    assert Enum.member?(res_headers, {"access-control-allow-origin", "*"})
+    # CORS should verify HTTP methods
+    assert Enum.member?(res_headers, {"access-control-allow-methods", "*"})
+    # CORS should verify allowed headers
+    assert Enum.member?(
+             res_headers,
+             {"access-control-allow-headers", "content-type,authorization"}
+           )
+  end
+
+  @tag :kafka
+  test "Given target is set to Kafka, the http request should publish message to Kafka topic" do
+    test_name = "proxy-http-response-publish-to-kafka"
+
+    api_id = "mock-#{test_name}-api"
+    endpoint_id = "mock-#{test_name}-endpoint"
+    endpoint_path = "/#{endpoint_id}"
+
+    # We register the endpoint with the proxy:
+    rig_api_url = "http://localhost:#{@api_port}/v1/apis"
+    rig_proxy_url = "http://localhost:#{@proxy_port}"
+
+    body =
+      Jason.encode!(%{
+        id: api_id,
+        name: "Mock API",
+        version_data: %{
+          default: %{
+            endpoints: [
+              %{
+                id: endpoint_id,
+                type: "http",
+                not_secured: true,
+                method: "POST",
+                path: endpoint_path,
+                target: "kafka"
+              }
+            ]
+          }
+        },
+        proxy: %{
+          target_url: "http://localhost",
+          port: 3000
+        }
+      })
+
+    headers = [{"content-type", "application/json"}]
+    HTTPoison.post!(rig_api_url, body, headers)
+
+    # The client calls the proxy endpoint:
+    request_url = rig_proxy_url <> endpoint_path
+
+    kafka_body =
+      Jason.encode!(%{
+        "data" => %{
+          "eventID" => "069711bf-3946-4661-984f-c667657b8d85",
+          "eventType" => "com.example.test",
+          "eventTypeVersion" => "1.0",
+          "eventTime" => "2018-04-05T17:31:00Z",
+          "cloudEventsVersion" => "0.1",
+          "source" => "/rig-test",
+          "contentType" => "application/json",
+          "extensions" => %{},
+          "data" => %{
+            "foo" => "bar"
+          }
+        },
+        "partition_key" => "test_key"
+      })
+
+    %HTTPoison.Response{status_code: res_status, body: res_body} =
+      HTTPoison.post!(request_url, kafka_body, headers)
+
+    assert res_status == 202
+    assert res_body == "Accepted."
+
+    expected_msg =
+      Jason.encode!(%{
+        "source" => "rig-test",
+        "rig" => %{
+          "scheme" => "http",
+          "requestPath" => "/mock-proxy-http-response-publish-to-kafka-endpoint",
+          "reqHeaders" => [
+            ["user-agent", "hackney/1.14.0"],
+            ["host", "localhost:4000"],
+            ["content-type", "application/json"],
+            ["content-length", "291"]
+          ],
+          "remoteIP" => "127.0.0.1",
+          "queryString" => "",
+          "port" => 4000,
+          "method" => "POST",
+          "host" => "localhost",
+          "correlationID" => "g2dkAA1ub25vZGVAbm9ob3N0AAADvgAAAAAA"
+        },
+        "extensions" => %{},
+        "eventTypeVersion" => "1.0",
+        "eventType" => "com.example.test",
+        "eventTime" => "2018-04-05T17:31:00Z",
+        "eventID" => "069711bf-3946-4661-984f-c667657b8d85",
+        "data" => %{
+          "foo" => "bar"
+        },
+        "contentType" => "application/json",
+        "cloudEventsVersion" => "0.1"
+      })
+
+    assert_receive expected_msg, 10_000
   end
 end